strategy-agent/pyproject.toml
[project]
name = "promote-autonomy-strategy-agent"
version = "0.1.0"
description = "AI strategy generation with HITL approval"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "google-cloud-firestore>=2.20.0",
    "google-cloud-pubsub>=2.26.0",
    "google-cloud-aiplatform>=1.75.0",
    "firebase-admin>=6.6.0",
    "python-dotenv>=1.0.0",
    "python-ulid>=3.0.0",
    "httpx>=0.28.0",
    "promote-autonomy-shared",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "black>=24.10.0",
    "mypy>=1.13.0",
    "types-requests>=2.32.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "black>=24.10.0",
    "mypy>=1.13.0",
]

[tool.uv.sources]
promote-autonomy-shared = { path = "../shared" }

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests requiring external services",
]
